Made SCA's iFirstAltVar & iLastAltVar to be short

+++ libpolys/polys/nc/nc.h
@@ -111,7 +111,7 @@ struct nc_struct
         // these variables should have odd degree, though that will not be checked
         // iAltVarsStart, iAltVarsEnd are only used together with nc_type=nc_exterior
         // 1 <= iAltVarsStart <= iAltVarsEnd <= r->N
-        unsigned int iFirstAltVar, iLastAltVar; // = 0 by default
+        short iFirstAltVar, iLastAltVar; // = 0 by default
 
         // for factors of super-commutative algebras we need 
         // the part of general quotient ideal modulo squares!    
@@ -124,14 +124,14 @@ struct nc_struct
     inline nc_type& ncRingType() { return (type); };
     inline nc_type ncRingType() const { return (type); };
 
-    inline unsigned int& FirstAltVar() 
+    inline short& FirstAltVar() 
         { assume(ncRingType() == nc_exterior); return (data.sca.iFirstAltVar); };
-    inline unsigned int& LastAltVar () 
+    inline short& LastAltVar () 
         { assume(ncRingType() == nc_exterior); return (data.sca.iLastAltVar ); };
 
-    inline unsigned int FirstAltVar() const 
+    inline short FirstAltVar() const 
         { assume(ncRingType() == nc_exterior); return (data.sca.iFirstAltVar); };
-    inline unsigned int LastAltVar () const 
+    inline short LastAltVar () const 
         { assume(ncRingType() == nc_exterior); return (data.sca.iLastAltVar ); };
 
     inline ideal& SCAQuotient()
